--- docEl.js
@@ -1,4 +1,0 @@
-export const btnKick = document.getElementById('btn-kick');
-export const kickEnemy = document.getElementById('kick-enemy');
-export const counterOne = document.querySelector('#btn-kick>.counter');
-export const counterTwo = document.querySelector('#kick-enemy>.counter');
